requirements.txt
alabaster==0.7.12
aniso8601==9.0.1
appdirs==1.4.4
argon2-cffi==21.1.0
asttokens==2.2.1
attrs==21.2.0
Babel==2.8.0
backcall==0.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==4.1.0
blinker==1.6.2
brial==1.0.2
Brotli==1.0.9
CacheControl==0.13.1
cachetools==5.3.1
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
colorama==0.4.4
comm==0.1.3
command-not-found==0.3
cryptography==3.4.8
cycler==0.11.0
cypari2==2.1.2
Cython==0.29.28
dbus-python==1.2.18
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.1
docutils==0.17.1
entrypoints==0.4
exceptiongroup==1.1.3
executing==1.2.0
filelock==3.6.0
firebase-admin==6.2.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
fonttools==4.29.1
fpylll==0.5.2.dev0
fs==2.4.12
gast==0.5.2
gevent==23.9.1
gmpy2==2.1.2
google-api-core==2.12.0
google-api-python-client==2.101.0
google-auth==2.23.2
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
greenlet==3.0.0
grequests==0.7.0
grpcio==1.59.0
grpcio-status==1.59.0
gyp==0.1
html5lib==1.1
httplib2==0.20.2
idna==3.3
imagesize==1.3.0
importlib-metadata==4.6.4
iniconfig==2.0.0
ipykernel==6.23.1
ipython==8.13.2
ipython_genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==2.1.2
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==3.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-pygments==0.1.2
keyring==23.5.0
kiwisolver==1.3.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==4.8.0
lz4==3.1.3+dfsg
MarkupSafe==2.1.3
matplotlib==3.5.1
matplotlib-inline==0.1.6
memory-allocator==0.1.2
more-itertools==8.10.0
mpmath==0.0.0
msgpack==1.0.7
nbclient==0.5.6
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.6
netifaces==0.11.0
networkx==2.4
notebook==6.4.8
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
packaging==23.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pkgconfig==1.5.5
platformdirs==3.5.1
pluggy==0.13.0
ply==3.11
pplpy==0.8.7
primecountpy==0.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.38
proto-plus==1.22.3
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
PyGObject==3.42.1
PyJWT==2.8.0
pyparsing==2.4.7
pyrsistent==0.18.1
pytest==7.4.2
pytest-asyncio==0.21.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
pythran==0.10.0
pytz==2022.1
PyYAML==5.4.1
pyzmq==25.1.0
requests==2.25.1
roman==3.3
rpy2==3.4.5
rsa==4.9
sagemath-standard==9.5
sagenb-export==3.2
scipy==1.8.0
SecretStorage==3.3.1
Send2Trash==1.8.1b0
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.3.1
Sphinx==4.3.2
stack-data==0.6.2
sympy==1.9
systemd-python==234
terminado==0.13.1
testpath==0.5.0
toml==0.10.2
tomli==2.0.1
tornado==6.3.2
tox==3.21.4
traitlets==5.9.0
tzlocal==2.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
uritemplate==4.1.1
urllib3==1.26.5
virtualenv==20.13.0+ds
wadllib==1.3.6
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==3.0.0
widgetsnbextension==2.0.0
xkit==0.0.0
zipp==1.0.0
zope.event==5.0
zope.interface==6.1
